[Tool] Support compilation on Mac 9: be code 3 (#63853)

+++ be/src/http/action/compaction_action.cpp
@@ -273,7 +273,8 @@ Status CompactionAction::_handle_show_repairs(HttpRequest* req, std::string* jso
         need_to_repair.PushBack(item, root.GetAllocator());
     }
     root.AddMember("need_to_repair", need_to_repair, root.GetAllocator());
-    root.AddMember("need_to_repair_num", tablets_with_small_segment_files.size(), root.GetAllocator());
+    root.AddMember("need_to_repair_num", static_cast<uint64_t>(tablets_with_small_segment_files.size()),
+                   root.GetAllocator());
 
     rapidjson::Document executed;
     executed.SetArray();
@@ -295,7 +296,7 @@ Status CompactionAction::_handle_show_repairs(HttpRequest* req, std::string* jso
         executed.PushBack(item, root.GetAllocator());
     }
     root.AddMember("executed_task", executed, root.GetAllocator());
-    root.AddMember("executed_task_num", tasks.size(), root.GetAllocator());
+    root.AddMember("executed_task_num", static_cast<uint64_t>(tasks.size()), root.GetAllocator());
 
     rapidjson::StringBuffer strbuf;
     rapidjson::PrettyWriter<rapidjson::StringBuffer> writer(strbuf);
